requirements.txt
# File: requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
h5py>=3.8.0
matplotlib>=3.7.0
tqdm>=4.65.0
pyyaml>=6.0
scikit-image>=0.20.0
tensorboard>=2.13.0
pytest>=7.3.0
pytest-cov>=4.1.0
fastmri>=0.3.0
einops>=0.6.0
wandb>=0.15.0
Pillow>=9.5.0